+++ nodes/utils/hueUtils.js
@@ -1,9 +1,7 @@
 'use strict'
 
-// const hueApi = require('node-hue-api')
-const hueApiV2 = require('node-hue')
 const { EventEmitter } = require('events')
-const https = require('https')
+const http = require('./http.js')
 const EventSource = require('eventsource');
 
 class classHUE extends EventEmitter {
@@ -20,11 +18,7 @@ class classHUE extends EventEmitter {
     this.commandQueue = []
     this.closePushEventStream = false
     this.timerwriteQueueAdd = setTimeout(this.handleQueue, 3000) // First start
-    this.connect()
-  }
 
-
-  connect() {
     // #############################################
     const options = {
       headers: {
@@ -61,18 +55,28 @@ class classHUE extends EventEmitter {
     }
     this.es.onerror = (error) => {
       try {
-        this.es.close()
-        this.es = null
-        console.log('KNXUltimateHUEConfig: classHUE: request.on(error): ' + error.message)
         setTimeout(() => {
           this.connect()
         }, 5000);
+        this.es.close()
+        this.es = null
+        console.log('KNXUltimateHUEConfig: classHUE: request.on(error): ' + error.message)
       } catch (error) {
-
       }
       //this.emit('error', err)
     };
+    // Initialize the http wrapper, to use the provided key.
+    // This http wrapper is used to get the data from HUE brigde
+    this.hueApiV2 = await http.use({
+      key: this.username,
+      prefix: 'https://' + this.hueBridgeIP + '/clip/v2'
+    });
+    // Load all resources, to avoid too many call to the HUE bridge and speed up the showing of the device mnames, during typing in the config window
+    this.hueAllResources = await this.hueApiV2.get('/resource')
+    this.hueAllRooms = await this.hueApiV2.get('/resource/room')
+    this.hueAllDevices = await this.hueApiV2.get('/resource/device')
   }
+
   // #############################################
 
 
@@ -83,35 +87,47 @@ class classHUE extends EventEmitter {
       const jRet = this.commandQueue.shift()
       switch (jRet._operation) {
         case 'setLight':
+          // It can be a light or a grouped light
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-            const ok = await hue.setLight(jRet._lightID, jRet._state)
+            const ok = await this.hueApiV2.put('/resource/light/' + jRet._lightID, jRet._state)
           } catch (error) {
-            console.log('KNXUltimateHUEConfig: classHUE: handleQueue: setLight: ' + error.message)
+            console.log('KNXUltimateHUEConfig: classHUE: handleQueue: setLight light: ' + error.message)
+          }
+          break
+        case 'setGroupedLight':
+          try {
+            const ok = await this.hueApiV2.put('/resource/grouped_light/' + jRet._lightID, jRet._state)
+          } catch (error) {
+            console.log('KNXUltimateHUEConfig: classHUE: handleQueue: setLight grouped_light: ' + error.message)
           }
           break
         case 'getLight':
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-            const jReturn = await hue.getLight(jRet._lightID)
+            const jReturn = await this.hueApiV2.get('/resource/light/' + jRet._lightID)
             jRet._callback(jReturn[0]) // Need to call the callback, because the event is absolutely async
           } catch (error) {
-            console.log('KNXUltimateHUEConfig: classHUE: handleQueue: getLight: ' + error.message)
+            console.log('KNXUltimateHUEConfig: classHUE: handleQueue: getLight light: ' + error.message)
+          }
+          break
+        case 'getGroupedLight':
+          try {
+            const jReturn = await this.hueApiV2.get('/resource/grouped_light/' + jRet._lightID)
+            jRet._callback(jReturn[0]) // Need to call the callback, because the event is absolutely async
+          } catch (error) {
+            console.log('KNXUltimateHUEConfig: classHUE: handleQueue: getLight grouped_light: ' + error.message)
           }
           break
         case 'setScene':
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
             const sceneID = jRet._lightID
-            const ok = await hue.setScene(sceneID, jRet._state)
+            const ok = await this.hueApiV2.put('/resource/scene/' + sceneID, jRet._state)
           } catch (error) {
             console.log('KNXUltimateHUEConfig: classHUE: handleQueue: setScene: ' + error.message)
           }
           break
         case 'stopScene':
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-            const allResources = await hue.getResources()
+            const allResources = await this.hueApiV2.get('/resource')
             const sceneID = jRet._lightID
             const jScene = allResources.find(res => res.id === sceneID) || ''
             const linkedLight = allResources.find(res => res.id === jScene.group.rid).children || ''
@@ -124,16 +140,14 @@ class classHUE extends EventEmitter {
           break
         case 'getBattery':
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-            const jReturn = await hue.getDevicePower(jRet._lightID)
+            const jReturn = await this.hueApiV2.get('/resource/device_power/' + jRet._lightID)
             jRet._callback(jReturn[0]) // Need to call the callback, because the event is absolutely async
           } catch (error) {
             console.log('KNXUltimateHUEConfig: classHUE: handleQueue: getBattery: ' + error.message)
           }
         case 'getLightLevel':
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-            const jReturn = await hue.getLightLevel(jRet._lightID)
+            const jReturn = await this.hueApiV2.get('/resource/light_level/' + jRet._lightID)
             jRet._callback(jReturn[0]) // Need to call the callback, because the event is absolutely async
           } catch (error) {
             console.log('KNXUltimateHUEConfig: classHUE: handleQueue: getLightLevel: ' + error.message)
@@ -141,8 +155,7 @@ class classHUE extends EventEmitter {
           break
         case 'getTemperature':
           try {
-            const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-            const jReturn = await hue.getTemperature(jRet._lightID)
+            const jReturn = await this.hueApiV2.get('/resource/temperature/' + jRet._lightID)
             jRet._callback(jReturn[0]) // Need to call the callback, because the event is absolutely async
           } catch (error) {
             console.log('KNXUltimateHUEConfig: classHUE: handleQueue: getTemperature: ' + error.message)
@@ -162,61 +175,90 @@ class classHUE extends EventEmitter {
   }
   // ######################################
 
+
+
   // Get all devices and join it with relative rooms, by adding the room name to the device name
   getResources = async (_rtype, _host, _username) => {
     try {
-      // V2
-      const hue = hueApiV2.connect({ host: _host, key: _username })
-      const retArray = []
-      const allResources = await hue.getResources()
-      const allRooms = await hue.getRooms()
-      const newArray = allResources.filter(x => x.type === _rtype)
+      // Api V2
 
-      newArray.forEach(jResource => {
-        if (_rtype === 'scene') {
-          const linkedZone = allResources.find(res => res.id === jResource.group.rid) || ''
-          retArray.push({ name: 'Scene: ' + jResource.metadata.name + (linkedZone !== undefined ? ', zone ' + linkedZone.metadata.name : ''), id: jResource.id })
-        } else {
-          const Room = allRooms.find(room => room.children.find(child => child.rid === jResource.owner.rid))
-          const linkedDevName = allResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === jResource.id)).metadata.name || ''
-          if (_rtype === 'button') {
-            const controlID = jResource.metadata !== undefined ? (jResource.metadata.control_id || '') : ''
-            retArray.push({ name: 'Button: ' + linkedDevName + (controlID !== '' ? ', button ' + controlID : '') + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+      // Returns capitalized string
+      function capStr(s) {
+        if (typeof s !== 'string') return ''
+        return s.charAt(0).toUpperCase() + s.slice(1)
+      }
+
+      const retArray = []
+      let allResources = undefined
+      if (_rtype === 'light' || _rtype === 'grouped_light')
+        allResources = await this.hueAllResources.filter(a => a.type === 'light' || a.type === 'grouped_light')
+      else {
+        allResources = await this.hueAllResources.filter(a => a.type === _rtype)
+      }
+      for (let index = 0; index < allResources.length; index++) {
+        const resource = allResources[index];
+        // Get the owner
+        try {
+          let resourceName = ''
+          let sRoom = ''
+          if (_rtype === 'light' || _rtype === 'grouped_light') {
+            // It's a service, having a owner
+            const owners = await this.hueAllResources.filter(a => a.id === resource.owner.rid)
+            for (let index = 0; index < owners.length; index++) {
+              const owner = owners[index];
+              //resourceName += (owner.metadata !== undefined && owner.metadata.name !== undefined) ? owner.metadata.name + ' and ' : ' and '
+              resourceName += owner.metadata.name + ' and '
+              const room = await this.hueAllRooms.find(child => child.children.find(a => a.rid === owner.id))
+              sRoom += room !== undefined ? room.metadata.name + ' + ' : ' + '
+            }
+            sRoom = sRoom.slice(0, -(' + '.length))
+            resourceName = resourceName.slice(0, -(' and '.length))
+            resourceName += sRoom !== '' ? ' - Room: ' + sRoom : ''
+            retArray.push({ name: capStr(resource.type) + ': ' + resourceName, id: resource.id })
           }
-          if (_rtype === 'light') {
-            retArray.push({ name: 'Light: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+          if (_rtype === 'scene') {
+            resourceName = resource.metadata.name || '**Name Not Found**'
+            // Get the linked zone
+            const zone = await this.hueAllResources.find(res => res.id === resource.group.rid)
+            resourceName += ' - ' + capStr(resource.group.rtype) + ': ' + zone.metadata.name
+            retArray.push({ name: capStr(_rtype) + ': ' + resourceName, id: resource.id })
+          }
+          if (_rtype === 'button') {
+            let linkedDevName = await this.hueAllResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === resource.id)).metadata.name || ''
+            const controlID = resource.metadata !== undefined ? (resource.metadata.control_id || '') : ''
+            retArray.push({ name: capStr(_rtype) + ': ' + linkedDevName + ', button ' + controlID, id: resource.id })
           }
           if (_rtype === 'motion') {
-            retArray.push({ name: 'Motion: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+            let linkedDevName = await this.hueAllResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === resource.id)).metadata.name || ''
+            retArray.push({ name: capStr(_rtype) + ': ' + linkedDevName, id: resource.id })
           }
           if (_rtype === 'relative_rotary') {
-            retArray.push({ name: 'Rotary: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+            let linkedDevName = await this.hueAllResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === resource.id)).metadata.name || ''
+            retArray.push({ name: 'Rotary: ' + linkedDevName, id: resource.id })
           }
           if (_rtype === 'light_level') {
-            retArray.push({ name: 'Light Level: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+            let Room = await this.hueAllRooms.find(room => room.children.find(child => child.rid === resource.owner.rid))
+            let linkedDevName = await this.hueAllResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === resource.id)).metadata.name || ''
+            retArray.push({ name: 'Light Level: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: resource.id })
           }
           if (_rtype === 'temperature') {
-            retArray.push({ name: 'Temperature: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+            let Room = await this.hueAllRooms.find(room => room.children.find(child => child.rid === resource.owner.rid))
+            let linkedDevName = await this.hueAllResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === resource.id)).metadata.name || ''
+            retArray.push({ name: 'Temperature: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: resource.id })
           }
           if (_rtype === 'device_power') {
-            retArray.push({ name: 'Battery: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: jResource.id })
+            let Room = await this.hueAllRooms.find(room => room.children.find(child => child.rid === resource.owner.rid))
+            let linkedDevName = await this.hueAllResources.find(dev => dev.type === 'device' && dev.services.find(serv => serv.rid === resource.id)).metadata.name || ''
+            retArray.push({ name: 'Battery: ' + linkedDevName + (Room !== undefined ? ', room ' + Room.metadata.name : ''), id: resource.id })
           }
+        } catch (error) {
+          //retArray.push({ name: _rtype + ': ERROR ' + error.message, id: resource.id })
         }
-      })
+      }
       return { devices: retArray }
     } catch (error) {
       console.log('KNXUltimateHue: HueUtils: classHUE: getDevices: error ' + error.message)
       return ({ devices: error.message })
-    }
-  }
-
-  // Get the light details
-  getLightStatus = async (_rid) => {
-    try {
-      const hue = hueApiV2.connect({ host: this.hueBridgeIP, key: this.username })
-      const oLight = await hue.getLight(_rid)
-      return oLight[0]
-    } catch (error) {
     }
   }